Redesign internal function __q_merge_two
Previous __q_merge_two takes actual head of the queue as input.
In this version, we takes the argument as a non circular linked list
and two head are the first entry of two queues.

In this design, function call to __q_merge_two. Does not need to
create a virtual head to pass linkd list in. Just pass the first
entry of the linked list and the function will work.